+++ app/src/main/java/eu/kanade/presentation/browse/components/BrowseBadges.kt
@@ -3,7 +3,9 @@ package eu.kanade.presentation.browse.components
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.CollectionsBookmark
 import androidx.compose.runtime.Composable
+import tachiyomi.i18n.MR
 import tachiyomi.presentation.core.components.Badge
+import tachiyomi.presentation.core.i18n.stringResource
 
 @Composable
 internal fun InLibraryBadge(enabled: Boolean) {
@@ -11,5 +13,16 @@ internal fun InLibraryBadge(enabled: Boolean) {
         Badge(
             imageVector = Icons.Outlined.CollectionsBookmark,
         )
+    }
+}
+
+@Composable
+internal fun BrowseLibraryBadge(
+    inLibrary: Boolean,
+    sameTitleLibraryMatch: Boolean,
+) {
+    when {
+        inLibrary -> InLibraryBadge(enabled = true)
+        sameTitleLibraryMatch -> Badge(text = stringResource(MR.strings.title_in_library))
     }
 }